requirements.txt
jupyter==1.1.1
pandas==2.3.0
numpy==2.2.6
tqdm==4.67.1
python-dotenv==1.1.1
transformers==4.53.1
torch==2.7.1
nltk==3.9.1
sentence-transformers==5.0.0
scikit-learn==1.7.0
scipy==1.15.3
matplotlib==3.10.3
lingua-language-detector==2.1.1
xgboost==3.0.2
google-genai==1.24.0
openai==1.93.0
thefuzz==0.22.1
spacy==3.8.7
# pyarabic==0.6.15
# camel-tools==1.5.6